Ajout de la possibilité de supprimer des articles dans le panier

+++ front/js/cart.js
@@ -1,35 +1,40 @@
 let panier = JSON.parse(localStorage.getItem("product"));
 let prixFinal = 0;
 let bool = false;
+let indice = 0
 
-for (article of panier) {
-  document.querySelector('#cart__items').innerHTML += `
-    <article class="cart__item" data-id="{product-ID}" data-color="{product-color}">
-                <div class="cart__item__img">
-                  <img src="${article.image}" alt="${article.alt}">
-                </div>
-                <div class="cart__item__content">
-                  <div class="cart__item__content__description">
-                    <h2>${article.name}</h2>
-                    <p>${article.color}</p>
-                    <p>${article.price} €</p>
+
+if (panier !== null) {
+  for (article of panier) {
+    document.querySelector('#cart__items').innerHTML += `
+      <article class="cart__item" data-id="{product-ID}" data-color="{product-color}">
+                  <div class="cart__item__img">
+                    <img src="${article.image}" alt="${article.alt}">
                   </div>
-                  <div class="cart__item__content__settings">
-                    <div class="cart__item__content__settings__quantity">
-                      <p>Qté : </p>
-                      <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${article.quantity}">
+                  <div class="cart__item__content">
+                    <div class="cart__item__content__description">
+                      <h2>${article.name}</h2>
+                      <p>${article.color}</p>
+                      <p>${article.price} €</p>
                     </div>
-                    <div class="cart__item__content__settings__delete">
-                      <p class="deleteItem">Supprimer</p>
+                    <div class="cart__item__content__settings">
+                      <div class="cart__item__content__settings__quantity">
+                        <p>Qté : </p>
+                        <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${article.quantity}">
+                      </div>
+                      <div class="cart__item__content__settings__delete">
+                        <p class="deleteItem">Supprimer</p>
+                      </div>
                     </div>
                   </div>
-                </div>
-              </article>`
-  prixFinal += article.price * article.quantity
+                </article>`
+    prixFinal += article.price * article.quantity
+  }
 
 }
 
 document.querySelector('#totalPrice').innerHTML = prixFinal;
+let buttonDelete = document.querySelectorAll('.deleteItem');
 
 let changeQuantity = document.querySelectorAll('.itemQuantity');
 
@@ -54,4 +59,34 @@ for (el of changeQuantity) {
       bool = false;
     }
   }))
+}
+let nouveauPanier = JSON.parse(localStorage.getItem("product"));
+
+for (el of buttonDelete) {
+  el.addEventListener('click', (e => {
+    let cibleName = e.target.parentElement.parentElement.parentElement.children[0].children[0].innerText;
+    let cibleColor = e.target.parentElement.parentElement.parentElement.children[0].children[1].innerText;
+    
+    panier = panier.map(element => {
+      if (element.name == cibleName && element.color == cibleColor) {
+        console.log(nouveauPanier);
+        indice = panier.indexOf(element)
+        console.log('voici lindice' + indice);
+        nouveauPanier.splice(indice, 1);
+        localStorage.setItem("product", JSON.stringify(nouveauPanier));
+        window.location.reload();
+        return panier;
+      } else {
+        return panier;
+      }
+      
+    })
+
+  }))
+}
+
+if (panier && panier.length == 0) {
+  console.log("supprimer");
+  localStorage.clear();
+  window.location.reload();
 }
